Create charge_hydrophobic_quantification
Input sequence as b (between the apostrophes) and run code. Results will show the total number of; charged and polar residues together,  hydrophobic residues and the total length of the sequence. It will also display the number. of methionines, tryptophan and any residues represented by X (which usually stands for any amino acid).